Assistant/services/agents/agents/config/llm_router/pyproject.toml
[project]
name = "llm-router"
version = "0.1.0"
description = "Multi-cloud LLM router with identity-based authentication"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "botocore>=1.35.0",
    "google-auth>=2.35.0",
    "google-cloud-aiplatform>=1.70.0",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.23.0",
    "pydantic-ai>=1.2",
    "certifi>=2024.8.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
